+++ packages/utils/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.1](https://github.com/noaignite/accelerator/compare/@noaignite/utils@1.1.0...@noaignite/utils@1.1.1) (2021-12-11)
+
+**Note:** Version bump only for package @noaignite/utils
+
+
+
+
+
+# 1.1.0 (2021-12-02)
+
+
+### Features
+
+* **utils:** add sleep util which resolves a promise after a specified delay ([10975b2](https://github.com/noaignite/oui/commit/10975b2bd14fbd8d98dc99ca65fa76257019f920))
+
+
+
+
+
 
 
 **Note:** Version bump only for package @noaignite/utils